+++ src/controllers/hotkeys/ActionNames.hpp
@@ -85,6 +85,7 @@ inline const std::map<HotkeyCategory, ActionDefinitionMap> actionNames{
           ActionDefinition{
               "Usercard: execute moderation action",
               "<ban, unban or number of the timeout button to use>", 1}},
+         {"pin", ActionDefinition{"Usercard, reply thread: pin window"}},
      }},
     {HotkeyCategory::Split,
      {